+++ Api/Services/Payments/IPaymentsService.cs
@@ -12,7 +12,5 @@ namespace HappyTravel.Edo.Api.Services.Payments
         Task<Result> ProcessPaymentChanges(Payment payment);
         
         Task<Result<(int AgentId, int AgencyId)>> GetServiceBuyer(string referenceCode);
-
-        Task<Result> SetPaymentMethodToCreditCard(string referenceCode);
     }
 }
